# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API WAFFER",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "https://nad-api.squaretech.tech/api"
        },
        {
            "url": "http://127.0.0.1:8000/api"
        }
    ],
    "paths": {
...
        "/admin/animal-species": {
            "get": {
                "tags": [
                    "Admin - Animal Species"
                ],
                "description": "Get all animal species",
                "operationId": "get_all_animal_species",
                "parameters": [
                    {
                        "name": "with_paginate",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "enum": [
                                0,
                                1
                            ]
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "category_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "animal_type_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "q",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin - Animal Species"
                ],
                "description": "Add new animal specie.",
                "operationId": "308fe8b8f1468ca36b49cff1469b5cc5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name[ar]",
                                    "category_id",
                                    "animal_type_id"
                                ],
                                "properties": {
                                    "category_id": {
                                        "type": "integer"
                                    },
                                    "animal_type_id": {
                                        "type": "integer"
                                    },
                                    "name[en]": {
                                        "type": "string"
                                    },
                                    "name[ar]": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/admin/animal-species/{id}": {
            "get": {
                "tags": [
                    "Admin - Animal Species"
                ],
                "description": "Get animal specie information.",
                "operationId": "show_animal_specie",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin - Animal Species"
                ],
                "description": "Edit animal specie.",
                "operationId": "bd5b1fb23c84acf5fe9c51aa2a30a7ef",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "category_id": {
                                        "type": "integer"
                                    },
                                    "animal_type_id": {
                                        "type": "integer"
                                    },
                                    "name[en]": {
                                        "type": "string"
                                    },
                                    "name[ar]": {
                                        "type": "string"
                                    },
                                    "_method": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "PUT"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin - Animal Species"
                ],
                "description": "Delete entered animal specie.",
                "operationId": "delete_animal_specie",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "int": {
                                "summary": "An int value.",
                                "value": "1"
                            }
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}